pyproject.toml
[project]
name = "daily-briefer"
version = "0.1.0"
description = "AI-powered daily briefing generator using crew.ai framework"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.65.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "crewai-tools[mcp]>=0.71.0",
]
